refactor: move status related files in a separate folder
